+++ dynamic-restserver/src/main/resources/com/microej/example/iot/script.js
@@ -1,9 +1,8 @@
 /*
  * Javascript
  *
- * Copyright 2019 MicroEJ Corp. All rights reserved.
+ * Copyright 2019-2023 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
- * MicroEJ Corp. PROPRIETARY. Use is subject to license terms.
  */
 
 const REFRESH_TIME = 1000