+++ CHANGELOG.rst
@@ -8,6 +8,15 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 -------------------
+16.0.0 - 2024-02-08
+-------------------
+
+Changed
+=======
+
+- Update to MicroEJ SDK 6 and switch from Ivy to Gradle.
+
+-------------------
 15.0.1 - 2023-01-10
 -------------------
 
@@ -155,6 +164,6 @@ Added
 --------------
 
 .. ReStructuredText
-.. Copyright 2016-2023 MicroEJ Corp. All rights reserved.
+.. Copyright 2016-2024 MicroEJ Corp. All rights reserved.
 .. Use of this source code is governed by a BSD-style license that can be found with this software.