+++ cbor/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2024-02-08
+
+### Changed
+
+  - Update to MicroEJ SDK 6 and switch from Ivy to Gradle.
+
 ## [2.0.0] - 2022-14-11
 
 ### Changed
@@ -22,5 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---  
 _Markdown_   
-_Copyright 2021-2022 MicroEJ Corp. All rights reserved._   
+_Copyright 2021-2024 MicroEJ Corp. All rights reserved._   
 _Use of this source code is governed by a BSD-style license that can be found with this software._