+++ dynamic-restserver/CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2024-02-08
+
+### Changed
+
+  - Update to MicroEJ SDK 6 and switch from Ivy to Gradle.
+  - Update license.
+
 ## [2.0.0] - 2022-14-11
 
 ### Changed
@@ -20,6 +27,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   
 ---  
 _Markdown_   
-_Copyright 2019-2022 MicroEJ Corp. All rights reserved._   
-_Use of this source code is governed by a BSD-style license that can be found with this software._   
-_MicroEJ Corp. PROPRIETARY. Use is subject to license terms._  
+_Copyright 2019-2024 MicroEJ Corp. All rights reserved._   
+_Use of this source code is governed by a BSD-style license that can be found with this software._